MPDX-3997 set a default location of other on person modal (#704)
* set a default location of other on person modal for email and phone

* spec fixes

+++ src/contacts/show/people/modal/phone/phone.component.js
@@ -1,10 +1,19 @@
+import isFunction from 'lodash/fp/isFunction';
+
 class ContactPhoneController {
     constructor() {
         this.deleted = false;
     }
+    $onInit() {
+        if (this.phone.location) {
+            this.phone.location = this.phone.location.toLowerCase();
+        } else {
+            this.phone.location = 'other';
+        }
+    }
     remove() {
         this.deleted = true;
-        this.onRemove();
+        if (isFunction(this.onRemove)) this.onRemove();
     }
 }
